Screen-space effects (#506)
@Josh-Schifter relayed a request from a user to implement a lens distortion post-processing effect to simulate the distortion produced by real-world cameras with very wide FOV. Rather than make such a specific effect part of imodeljs-frontend, I added a way for any app to apply custom post-processing effects. They define a shader program that takes as input the image rendered by the Viewport and outputs a modified image. This initial implementation is intentionally rather limited: it doesn't support additional input textures, intermediate rendering steps, etc - but it will serve this user's purpose.

Added a lens distortion effect to frontend-devtools to serve as an example for the user.
Also added various convolution effects like emboss, sharpen, and blur as more standard examples; and a dumb but simple image-flipping effect to serve as an easy-to-digest example.

Effects like lens distortion that shift pixels around instead of simply recoloring them will work with element locate, but not with snapping - there's no linear transform that can invert the distortion applied by the shader. If we really need to, we could allow the caller to specify a javascript function that would perform the equivalent of the inverse of the shader's effect so that we can at least obtain the original world position.

Each Viewport has a list of effects to be applied to it. The effects are applied in the order in which they appear in the list. They are applied after *everything else* has been drawn, so things like decorations and hilite silhouettes are also affected.