Improve error thrown when deserializing invalid JSON. (#588)
Currently, if an RPC value contains invalid JSON, we just get an SyntaxError with a message like _"Unexpected token T in JSON at position 0"_.  This is not usually an issue, since we never expect backends to respond with malformed JSON (even in error responses).  However, this issue can arise when gateways or proxys sit between a frontend and backend, and when that does happen, including the entire malformed JSON in the error message is much more helpful.