UI: Fix calling `scrollToNode` too early triggering an assertion error (#689)
If `scrollToNode` is called within `useEffect` that is run immediately after `TreeRenderer` component is mounted, an assertion error occurs. This is due to contents of `TreeRenderer` being rendered in two passes: first it determines its container width and height, and then it mounts `VariableSizeList` on the second pass.

This change introduces a new internal component called `VirtualizedList` which wraps `VariableSizeList` and adds the expected behaviour regarding `scrollToNode`. I don't think this is the ideal solution, the better one is to move `AutoSizer` component outside and let `TreeRenderer` consumers to specify tree width and height, but that would be a breaking change.