Support the existing environment variable. Disable failing test (#1028)
The update to use a baseUrl vs. Buddi we were missing the backwards compatibility with the imjs_buddi_resolve_url_using_region environment variable. Add support for it in the case a imjs_url_prefix is not defined.

There is an issue with the MayLayers usage of product settings service that needs to be investigated. Disabling for now until it can be resolved.