Remove the ulas utilities in backend (#1921)
The usage logging utilities are not available for 3rd parties to use and are a Bentley specific implementation detail. They are being moved out of the native addon and going to be handled internally.

The main consequence of this change, besides the removal of the APIs, is the openForRead call will no longer automatically log usage. We are working internally to define another strategy with the licensing team and various Bentley teams depending on this functionality which will be announced (internally) once it's ready.