NOTE COMMIT - I seem to have encountered a bug
after inserting a model selector with model props, it is immediately
viewable. However, upon the corresponding aspect insert (really any
insert), the models property is no longer visible... I assume it has
something to do with caching and the element being in memory during the
previous insert. Either way it is a stopper on this work for now