Fix jumpy view zoom on touch devices (#3498)
* Remove inertia from view touch handling as soon as zoom is detected

Previously required another event for inertia to be removed and not
re-added. If touches ended before then, inertia would cause the view
to incorrectly pan after a zoom gesture.

* Properly handle touch count change in

Previously if a touch was removed and re-added without a touch
move event in between, the change in touch count would go
undetected. In pinch-lift one finger-pinch scenario that caused
zoom to be calculated from first pinch start position, causing zoom
jumps at the start of the second pinch.

* rush change

* rush extract-api