Rename transformer predecessors to references (#3858)
Since 3.2.0 when the transformer started handling cyclic and backwards references, the term `predecessors` became inaccurate and confusing, since it implied that an element was definitely inserted before another, which is what the transformer used to only be able to handle (with some exceptions due to the old "deferral" mechanism). `getPredecessorIds`/`collectPredecessorIds` never really returned predecessors but references, and now we have `Element.requiredReferences` which returns real predecessors, adding to the confusion.

This PR deprecates (for removal in 4.x) `Element.getPredecessorIds` and `Element.collectPredecessorIds` and replaces them with `Element.getReferenceIds` and `Element.collectReferenceIds` to match the correct vocabulary. It also removes a lot of internal references to the "predecessor" vocabulary. 

Co-authored-by: Michael Belousov <MichaelBelousov@users.noreply.github.com>