Remove @iTwin/itwinjs-core team as owner for all changelog files.
I added this line in #7292 with the hopes of not requiring reviews from everyone for package.json-only updates.  An unintentional side-effect of this has been that now @iTwin/itwinjs-core is required to review every single PR, even when a smaller set of codeowners would otherwise be sufficient.

Since changelog files are never expected to be added on their own, let's just not require anyone as a reviewer for those.